Какая-то первая авторизация и регистрация

+++ database/migrations/2025_12_24_143541_create_personal_access_tokens_table.php
@@ -11,6 +11,8 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('personal_access_tokens')) return;
+
         Schema::create('personal_access_tokens', function (Blueprint $table) {
             $table->id();
             $table->morphs('tokenable');
@@ -28,6 +30,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (!Schema::hasTable('personal_access_tokens')) return;
+
         Schema::dropIfExists('personal_access_tokens');
     }
 };
